Wired up dataset creation to the service layer with some unit tests.
Note that this patient is still lying open on the table - much refactoring to do to avoid code duplication in the unit tests and cleaner access to the DAOImpls.

Also note that the Message and Comment model objects will either be svn-deleted soon or converted over to the DAO pattern.  They were merely example model objects to get the service implementation going.