Update
SWC startAndTrack treats all asynchronous jobs the same. However, to be
able to track some jobs, we need to know if it has an entityId. To
solve this, we add another interface called “HasEntityId.” Any
RequestBody that implements this interface must have an entityId. It’s
important that we generate the right url that contains the entityId to
track the job.