Update SynapseClient
Now update SynapseClient to use the new getStartUrl() and
getResultUrl().

Remove tableId from startAsynchJob and getAsyncResult since it is
included in the request body.

Overload getAsyncResult to take an entityId. This method helps with
tests.

Also, remove tableId as a parameter of uploadCsvTablePreviewAsyncStart
and uploadCsvToTablePreviewAsyncGet.

Finally, update integration test to use the new methods.