PLFM-6764: Remove offset from condition
We decided to avoid using pagination for this, the rationale behind is because the worker that will consume the messages for each key might be faster than the job driving it making the pagination unpredictable. Instead we adopt a simple approach with a bigger batch size since the job won't run very often and will eventually catch up. We decided to exclude alternatives such a 2 phase commit for now to process bigger batches since we do not expect to have millions of unlinked file handles every month.