PLFM-6766: Remove RESTORING status for file handles
While I was initially thinking of using S3 notifications on the restore complete to set the file handles back to AVAILABLE I realized that it would be too complex to setup an integration test. Additionally we would have to implement a recovery worker that scans the RESTORING file handles to account for missing messages. This is a lot of engineering for a service that we hopefully will use very rarely. I decided to simply submit the restore request and set the file handle as AVAILABLE instead. Users will have to wait a few hours before they can actually download those file handles while they are restored by S3.