PLFM-7017: Necessary refactor due to how entity service works
I initially wanted to use the versionNumber and isLatestVersion of the materializedView object but it turns out that the entity service does not provide the object read from the DB to the "afterCreate/Update" methods but rather what the client supplied that might not have those property or worse they might be wrong. Changing it to actually sync from the DB before the operation, brakes the rest of the code because the of how it is structured, the schema of a table is stored in a different table in the afterCreate/update from the original columnIds list in the object: if we read from the entity from the DB before the afterCreate/update the columns ids are not present anymore since they are read from the companion table.